mostly working launcher functions
