+++ run_rivers2stratigraphy.py
@@ -1,7 +1,9 @@
-import os
+import rivers2stratigraphy
 
-## initialize the GUI
-thisDir = os.path.dirname(__file__)
-thisPath = os.path.join(thisDir,'')
-execFile = os.path.join(thisPath, 'rivers2stratigraphy', 'rivers2stratigraphy.py')
-exec(open(execFile).read())
+# import os
+
+# ## initialize the GUI
+# thisDir = os.path.dirname(__file__)
+# thisPath = os.path.join(thisDir,'')
+# execFile = os.path.join(thisPath, 'rivers2stratigraphy', 'rivers2stratigraphy.py')
+# exec(open(execFile).read())
